Allowing appending additional nodes to the Excel file. Adapting GUI to have a proper view of the topologies. Adding support to ring structures apart from metro mesh.

+++ network.py
@@ -4,6 +4,8 @@ import networkx as nx
 import random
 from scipy.spatial import distance
 import pandas as pd
+from openpyxl import load_workbook
+import os.path
 
 
 # Generate topology with the specific degrees and weights and the specified number of nodes
@@ -19,13 +21,13 @@ def gen_topology(degrees, weights, nodes):
         updated = False
         # TODO Keep the smallest degree just in case there is not any with degrees[0]
         for i in range(len(sequence)):
-            # Look for the first element with smallest degree value and increase it, then break
+            # Look for the first element with the smallest degree value and increase it, then break
             if sequence[i] == degrees[0]:
                 # print("Incremented " + str(i))
                 sequence[i] += 1
                 updated = True
                 break
-            elif min_degree>sequence[i]:
+            elif min_degree > sequence[i]:
                 min_degree = sequence[i]
                 min_degree_idx = i
         if not updated:
@@ -62,7 +64,7 @@ def calculate_edge_distances(graph, positions, max_limit):
 # Rename nodes combining Type and index
 def rename_nodes(assigned_types, types, special_rename_type=None, explicit_values=None):
     indexes = dict(zip(types.code, [1] * len(types)))
-    print(indexes)
+    # print(indexes)
     # names = [''] * len(assigned_types)
     # for i in range(len(assigned_types)):
     #    names[i] = assigned_types[i] + str(indexes[assigned_types[i]])
@@ -75,52 +77,67 @@ def rename_nodes(assigned_types, types, special_rename_type=None, explicit_value
 # Rename nodes combining Type and index and receiving initial index per type
 def rename_nodes_idx(assigned_types, indexes, special_rename_type=None, explicit_values=None):
     names = [''] * len(assigned_types)
-    print(explicit_values)
+    # print(explicit_values)
     for i in range(len(assigned_types)):
         if assigned_types[i] == special_rename_type:
-            names[i] = explicit_values[indexes[assigned_types[i]]-1]
+            names[i] = explicit_values[indexes[assigned_types[i]] - 1]
         else:
             names[i] = assigned_types[i] + str(indexes[assigned_types[i]])
         indexes[assigned_types[i]] += 1
     return names, indexes
 
 
-# Write excel file with two sheets
-def write_network_xls(file, graph, distances, types, node_sheet, link_sheet, macro_region=None, x_coord=None, y_coord=None):
+# Write Excel file with two sheets
+def write_network_xls(file, graph, distances, types, node_sheet, link_sheet, macro_region=None, x_coord=None,
+                      y_coord=None, reference_nodes=None):
     if macro_region is None:
-        macro_region = [1]*len(graph.nodes)
+        macro_region = [1] * len(graph.nodes)
     if x_coord is None:
-        x_coord = [0]*len(graph.nodes)
+        x_coord = [0] * len(graph.nodes)
     if y_coord is None:
-        y_coord = [0]*len(graph.nodes)
+        y_coord = [0] * len(graph.nodes)
+    if reference_nodes is None:
+        reference_nodes = [''] * len(graph.nodes)
     nodes_df = pd.DataFrame({'node_name': [val for val in graph.nodes],
-                             'node_code': ['HL1']*len(graph.nodes),
-                             'Node Type': ['Backbone']*len(graph.nodes),
+                             'node_code': ['HL1'] * len(graph.nodes),
+                             'Node Type': ['Backbone'] * len(graph.nodes),
                              'Central office type': types,
                              'Reference Regional CO': [val for val in graph.nodes],
-                             'Reference National CO': ['']*len(graph.nodes),
-                             'Households':[12000]*len(graph.nodes),
-                             'Macro cells sites':[8]*len(graph.nodes),
-                             'Small cell sites': [24]*len(graph.nodes),
-                             'Twin Regional CO': ['']*len(graph.nodes),
-                             'Twin National CO': ['']*len(graph.nodes),
+                             'Reference National CO': reference_nodes,
+                             'Households': [12000] * len(graph.nodes),
+                             'Macro cells sites': [8] * len(graph.nodes),
+                             'Small cell sites': [24] * len(graph.nodes),
+                             'Twin Regional CO': [''] * len(graph.nodes),
+                             'Twin National CO': [''] * len(graph.nodes),
                              'Macro region': macro_region,
                              'x coord': x_coord,
                              'y_coord': y_coord
-                             #'Degrees': [val for u, val in graph.degree],
+                             # 'Degrees': [val for u, val in graph.degree],
                              })
     links_df = pd.DataFrame({'sourceID': [u for u, v in graph.edges],
                              'destinationID': [v for u, v in graph.edges],
                              'distanceKm': distances,
-                             'capacityGbps': [100]*len(graph.edges)}
+                             'capacityGbps': [100] * len(graph.edges)}
                             )
-    with pd.ExcelWriter(file, engine='openpyxl') as writer:
-        nodes_df.to_excel(writer, sheet_name=node_sheet, index=False)
-        links_df.to_excel(writer, sheet_name=link_sheet, index=False)
+    if os.path.isfile(file):
+        with pd.ExcelWriter(file, engine='openpyxl', mode="a", if_sheet_exists='overlay') as writer:
+            book=load_workbook(file)
+            # writer.book=book
+            writer.sheets.update(dict((ws.title, ws) for ws in book.worksheets))
+
+            nodes_df.to_excel(writer, sheet_name=node_sheet, index=False, header=False,
+                              startrow=writer.sheets[node_sheet].max_row)
+            links_df.to_excel(writer, sheet_name=link_sheet, index=False, header=False,
+                              startrow=writer.sheets[link_sheet].max_row)
+    else:
+        with pd.ExcelWriter(file, engine='openpyxl') as writer:
+            nodes_df.to_excel(writer, sheet_name=node_sheet, index=False)
+            links_df.to_excel(writer, sheet_name=link_sheet, index=False)
+
     return True
 
 
-# Write excel file with two sheets
+# Write Excel file with two sheets
 def write_network_xls_plus(file, graph, distances, types, node_sheet, link_sheet, reference_nodes):
     nodes_df = pd.DataFrame({'Nodes': [val for val in graph.nodes],
                              'Degrees': [val for u, val in graph.degree],
@@ -159,7 +176,7 @@ def backbone_assign_types(graph, types):
 
 # Assign types to graph nodes based on type proportions
 def metro_assign_types(types):
-    print("metro_assign_types: ", types)
+    # print("metro_assign_types: ", types)
     sequence = [code for code, val in types[['code', 'number']].values for _ in range(val)]
     # Generate k random values with the weight frequencies of degrees
     random.shuffle(sequence)
@@ -189,10 +206,10 @@ def match_sub_metro_region(topo_nodes, assigned_types, sub_topo_nodes, sub_topo_
     type_1 = sub_topo_types[0]
     idx_1 = assigned_types.index(type_1)
     type_2 = sub_topo_types[1]
-    idx_2 = assigned_types.index(type_2, idx_1+1)
+    idx_2 = assigned_types.index(type_2, idx_1 + 1)
     result = [idx_1, idx_2]
     result.extend(relabelled[2:(len(relabelled))])
-    print(result)
+    # print(result)
     return result, sub_topo_types[2:]
 
 
@@ -210,7 +227,7 @@ def connect_sub_metro_regions(topology, sub_topology, min_degree):
     topo_cons = random.sample(indexes, k=2)
     relabelled = [val + len(topology.nodes) for val in sub_topology.nodes]
 
-    indexes =[]
+    indexes = []
     while True:
         temp = [index + len(topology.nodes) for index, element in sub_topology.degree if element == low_degree]
         indexes.extend(temp)
@@ -221,7 +238,7 @@ def connect_sub_metro_regions(topology, sub_topology, min_degree):
 
     sub_topology = nx.relabel_nodes(sub_topology, dict(zip(sub_topology.nodes, relabelled)))
     topology = nx.union(topology, sub_topology)
-    topology.add_edge(topo_cons[0],subtopo_cons[0])
+    topology.add_edge(topo_cons[0], subtopo_cons[0])
     topology.add_edge(topo_cons[1], subtopo_cons[1])
     return topology
 
@@ -233,4 +250,3 @@ def format_distance_limits(distances, upper_limits):
         text += (str(init_dist) + "-" + str(upper) + ":\t" + str(perc) + "%\n")
         init_dist = upper
     return text
-
